requirements.txt
python-binance==1.0.19
pandas==2.0.2
numpy==1.24.3
scikit-learn==1.2.2
PyYAML==6.0.1
requests==2.28.2
ta-lib==0.4.20
matplotlib==3.7.1
tensorflow==2.12.0